delete confirm added

+++ src/pages/owner/index.jsx
@@ -106,13 +106,18 @@ const OwnerDashboard = () => {
   };
 
   const deleteController = async (id) => {
-    try {
-      await request.delete(`/auth/delete-account/${id}`);
-      getData();
-      message.success(`Deleted successfully`);
-    } catch (error) {
-      console.error("Error deleting data:", error);
-      message.error("Error deleting data");
+    const canIDelete = confirm("O'chirishga ishongching komilmi?");
+    if (canIDelete) {
+      try {
+        await request.delete(`/auth/delete-account/${id}`);
+        getData();
+        message.success(`O'chirildi`);
+      } catch (error) {
+        console.error("Error deleting data:", error);
+        message.error("Error deleting data");
+      }
+    } else {
+      message.error("O'chirish bekor qilindi");
     }
   };
